Replace unsigned with int

+++ cpp/grid.cpp
@@ -1,15 +1,15 @@
 #include "grid.hpp"
 
 Grid::Grid() : Grid(1, 1) {}
-Grid::Grid(unsigned dim, unsigned char colors) : Grid(dim, dim, colors) {}
-Grid::Grid(unsigned w, unsigned h, unsigned char colors) {
+Grid::Grid(int dim, unsigned char colors) : Grid(dim, dim, colors) {}
+Grid::Grid(int w, int h, unsigned char colors) {
     this->w = w;
     this->h = h;
     this->size = w * h;
     this->graph = new unsigned char[size];
     std::fill(graph, graph + this->size, 0);
     this->colors = colors;
-    this->counts = new unsigned[colors];
+    this->counts = new int[colors];
     std::fill(counts, counts + this->colors, 0);
     this->counts[0] = this->size;
 }
@@ -19,7 +19,7 @@ Grid::~Grid() {
     delete[] counts;
 }
 
-void Grid::set(unsigned index, unsigned char new_color) {
+void Grid::set(int index, unsigned char new_color) {
     unsigned char old_color = this->graph[index];
     if (old_color != new_color) {
         this->counts[old_color]--;
@@ -36,8 +36,8 @@ void Grid::set_all(unsigned char color) {
 
 void Grid::chessboard() {
     std::fill(counts, counts + this->colors, 0);
-    for (unsigned i = 0; i < this->h; i++) {
-        for (unsigned j = 0; j < this->w; j++) {
+    for (int i = 0; i < this->h; i++) {
+        for (int j = 0; j < this->w; j++) {
             this->graph[(i * this->w) + j] = (i + j) % this->colors;
             this->counts[(i + j) % this->colors]++;
         }
@@ -46,7 +46,7 @@ void Grid::chessboard() {
 
 void Grid::rand() {
     std::fill(counts, counts + this->colors, 0);
-    for (unsigned i = 0; i < size; i++) {
+    for (int i = 0; i < size; i++) {
         graph[i] = choose_color(colors);
         counts[graph[i]]++;
     }
@@ -57,8 +57,8 @@ void Grid::print() const {
 }
 
 void Grid::print(std::ostream& os) const {
-    for (unsigned i = 0; i < this->size; i++) {
-        os << (unsigned)this->graph[i] << " ";
+    for (int i = 0; i < this->size; i++) {
+        os << this->graph[i] << " ";
         if (i % this->w == this->w - 1)
             os << "\n";
     }
@@ -71,7 +71,7 @@ void Grid::print_counts() const {
 
 void Grid::print_counts(std::ostream& os) const {
     os << "Counts: ";
-    for (unsigned i = 0; i < this->colors; i++) {
+    for (int i = 0; i < this->colors; i++) {
         os << counts[i] << " ";
     }
     os << "\n";
@@ -85,9 +85,9 @@ Grid& Grid::operator=(const Grid& other) {
     delete[] this->graph;
     delete[] this->counts;
     this->graph = new unsigned char[this->size];
-    this->counts = new unsigned[this->colors];
+    this->counts = new int[this->colors];
     memcpy(this->graph, other.graph, sizeof(unsigned char) * this->size);
-    memcpy(this->counts, other.counts, sizeof(unsigned) * this->colors);
+    memcpy(this->counts, other.counts, sizeof(int) * this->colors);
 
     return *this;
 }
@@ -95,7 +95,7 @@ Grid& Grid::operator=(const Grid& other) {
 bool Grid::operator==(const Grid& other) {
     if (this->w != other.w || this->h != other.w) return false;
     if (this->colors != other.colors) return false;
-    for (unsigned i = 0; i < this->size; i++) {
+    for (int i = 0; i < this->size; i++) {
         if (this->graph[i] != other.graph[i])
             return false;
     }
@@ -113,32 +113,32 @@ void sum_neighbors(const Grid& g, int index, unsigned counts[]) {
     counts[g.graph[mod(index+1, g.size)]]++;
 }
 
-unsigned choose_point(const Grid& g) {
-    return (unsigned)(uniform(generator) * g.size);
+int choose_point(const Grid& g) {
+    return (int)(uniform(generator) * g.size);
 }
 
-void print_array(const Grid grids[], unsigned size) {
+void print_array(const Grid grids[], int size) {
     print_array(std::cout, grids, size);
 }
 
-void print_array(const Grid** grids, unsigned size) {
+void print_array(const Grid** grids, int size) {
     print_array(std::cout, grids, size);
 }
 
-void print_array(std::ostream& os, const Grid** grids, unsigned size) {
-    unsigned max_w = 0, max_h = 0, grid_w, grid_h;
-    for (unsigned i = 0; i < size; i++) {
+void print_array(std::ostream& os, const Grid** grids, int size) {
+    int max_w = 0, max_h = 0, grid_w, grid_h;
+    for (int i = 0; i < size; i++) {
         if (grids[i]->w > max_w) max_w = grids[i]->w;
         if (grids[i]->w > max_h) max_h = grids[i]->h;
     }
 
-    for (unsigned h = 0; h < max_h; h++) {
-        for (unsigned i = 0; i < size; i++) {
+    for (int h = 0; h < max_h; h++) {
+        for (int i = 0; i < size; i++) {
             grid_w = grids[i]->w;
             grid_h = grids[i]->h;
-            for (unsigned w = 0; w < max_w; w++) {
+            for (int w = 0; w < max_w; w++) {
                 if (w < grid_w && h < grid_h) {
-                    os << (unsigned)grids[i]->graph[h * grid_w + w] << ' ';
+                    os << grids[i]->graph[h * grid_w + w] << ' ';
                 }
             }
             os << ' ';
@@ -146,25 +146,25 @@ void print_array(std::ostream& os, const Grid** grids, unsigned size) {
         os << '\n';
     }
 
-    for (unsigned i = 0; i < size; i++) {
+    for (int i = 0; i < size; i++) {
         grids[i]->print_counts();
     }
 }
 
-void print_array(std::ostream& os, const Grid grids[], unsigned size) {
-    unsigned max_w = 0, max_h = 0, grid_w, grid_h;
-    for (unsigned i = 0; i < size; i++) {
+void print_array(std::ostream& os, const Grid grids[], int size) {
+    int max_w = 0, max_h = 0, grid_w, grid_h;
+    for (int i = 0; i < size; i++) {
         if (grids[i].w > max_w) max_w = grids[i].w;
         if (grids[i].w > max_h) max_h = grids[i].h;
     }
 
-    for (unsigned h = 0; h < max_h; h++) {
-        for (unsigned i = 0; i < size; i++) {
+    for (int h = 0; h < max_h; h++) {
+        for (int i = 0; i < size; i++) {
             grid_w = grids[i].w;
             grid_h = grids[i].h;
-            for (unsigned w = 0; w < max_w; w++) {
+            for (int w = 0; w < max_w; w++) {
                 if (w < grid_w && h < grid_h) {
-                    os << (unsigned)grids[i].graph[h * grid_w + w] << ' ';
+                    os << grids[i].graph[h * grid_w + w] << ' ';
                 }
             }
             os << ' ';
@@ -172,7 +172,7 @@ void print_array(std::ostream& os, const Grid grids[], unsigned size) {
         os << '\n';
     }
 
-    for (unsigned i = 0; i < size; i++) {
+    for (int i = 0; i < size; i++) {
         grids[i].print_counts();
     }
 }
@@ -181,15 +181,15 @@ bool mag_diff(const Grid& g1, const Grid& g2) {
     if (g1.colors != g2.colors)
         return true;
 
-    unsigned* counts1 = new unsigned[g1.colors];
-    unsigned* counts2 = new unsigned[g2.colors];
-    memcpy(counts1, g1.counts, sizeof(unsigned) * g1.colors);
-    memcpy(counts2, g2.counts, sizeof(unsigned) * g2.colors);
+    int* counts1 = new int[g1.colors];
+    int* counts2 = new int[g2.colors];
+    memcpy(counts1, g1.counts, sizeof(int) * g1.colors);
+    memcpy(counts2, g2.counts, sizeof(int) * g2.colors);
 
     std::sort(counts1, counts1 + g1.colors);
     std::sort(counts2, counts2 + g2.colors);
 
-    for (unsigned c = 0; c < g1.colors; c++) {
+    for (int c = 0; c < g1.colors; c++) {
         if (abs(counts1[c] - counts2[c]) > sqrt(g1.size / g1.colors)) {
             delete[] counts1;
             delete[] counts2;
